refactor/#46 : 코드리뷰 사항을 반영한다
