+++ frontend/src/components/frame/workspace/MainNavigationSidebar.tsx
@@ -1,7 +1,6 @@
 import WorkspaceIconButton from '@/components/workspace/WorkspaceIconButton';
 import { NAVIGATION_ICONS } from '@/constants/navItems';
-import { useWorkspaceSidebarStore } from '@/stores/workspace.store';
-import { SidebarType } from '@/stores/workspace.store';
+import { SidebarType, useWorkspaceSidebarStore } from '@/stores/workspace';
 
 const MainNavigationSidebar = () => {
   const icons = Object.values(NAVIGATION_ICONS);
@@ -9,9 +8,10 @@ const MainNavigationSidebar = () => {
 
   return (
     <div className=" min-w-16 bg-yellow-200 text-white flex py-3 flex-col items-center border-r-2">
-      {icons.map(item => {
+      {icons.map((item, index) => {
         return (
           <WorkspaceIconButton
+            key={index}
             className="bg-transparent"
             onClick={() => {
               if (item.type === 'ETC') console.log('더보기 클릭');
