+++ frontend/src/components/frame/workspace/SidebarContainer.tsx
@@ -1,6 +1,6 @@
 import WorkspaceContentSidebar from '@/components/frame/workspace/WorkspaceContentSidebar';
 import MainNavigationSidebar from '@/components/frame/workspace/MainNavigationSidebar';
-import currentFrameState from '@/stores/currentFrameState.store';
+import currentFrameState from '@/stores/currentFrameState';
 
 const SideBarContainer = () => {
   const { currentPage } = currentFrameState();
