+++ frontend/src/components/workspace/WorkspaceCreationProcess/index.tsx
@@ -1,11 +1,11 @@
+import { JSX } from 'react/jsx-runtime';
+import { useEffect } from 'react';
 import StepSetWorkspaceName from '@/components/workspace/WorkspaceCreationProcess/StepSetWorkspaceName';
 import StepSetUserName from '@/components/workspace/WorkspaceCreationProcess/StepSetUserName';
 import StepSetInviteUsers from '@/components/workspace/WorkspaceCreationProcess/StepSetInviteUsers';
-import { useWorkspaceCreationStore } from '@/stores/workspace.store';
-import { JSX } from 'react/jsx-runtime';
 import { stepData } from '@/lib/workspace';
-import currentFrameState from '@/stores/currentFrameState.store';
-import { useEffect } from 'react';
+import { useWorkspaceCreationStore } from '@/stores/workspace';
+import currentFrameState from '@/stores/currentFrameState';
 
 const WorkspaceCreationProcess = () => {
   const { step, workspaceName } = useWorkspaceCreationStore();
