+++ frontend/src/components/workspace/WorkspaceDashboard/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useParams } from 'react-router';
-import currentFrameState from '@/stores/currentFrameState.store';
+import currentFrameState from '@/stores/currentFrameState';
 
 const WorkspaceDashboard = () => {
   const { setCurrentPage } = currentFrameState();
